refactor : worldcupList component 합성형태로 변경
