+++ front/src/components/WorldCupItem/index.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { FaPlay, FaList, FaShare } from 'react-icons/fa';
 import { Link } from 'react-router-dom';
-import ShareModal from '../../ShareModal';
-import Image from '../../Image';
+import ShareModal from '../ShareModal';
+import Image from '../Image';
 
 interface Props {
   id: number;
