+++ front/src/pages/MyWorldcup/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Header from '../../components/Header';
+import MyWorldCupItem from '../../components/MyWorldCupItem';
 import { WorldcupList } from '../../components';
 import { getMyWorldcupList } from '../../apis/worldcups';
 import { useInfiniteScroll } from '../../hooks';
@@ -18,13 +19,22 @@ const MyWorldcup = (): JSX.Element => {
     <>
       <Header />
       <WorldcupList
-        type="myWorldcup"
-        worldcups={worldcups}
         observeTarget={target}
         isLoading={isLoading}
         isClickMore={isClickMore}
         onClickMoreBtn={onClickMoreBtn}
-      />
+      >
+        {worldcups.map(({ id, thumbnail1, thumbnail2, title, description }) => (
+          <MyWorldCupItem
+            key={id}
+            id={id}
+            thumbnail1={thumbnail1}
+            thumbnail2={thumbnail2}
+            title={title}
+            desc={description}
+          />
+        ))}
+      </WorldcupList>
     </>
   );
 };
